suggestion menu

+++ src/main/java/com/kobe/warehouse/constant/EntityConstant.java
@@ -10,7 +10,7 @@ public final class EntityConstant {
     public static final long SANS_EMPLACEMENT = 1;
     public static final int DEFAULT_STORAGE = 1;
     public static final String DEFAULT_MAIN_STORAGE = "DEFAULT_MAIN_STORAGE";
-    public static final long DEFAULT_MAGASIN = 1;
+    public static final int DEFAULT_MAGASIN = 1;
     public static final String APP_GESTION_STOCK = "APP_GESTION_STOCK";
     public static final String APP_MONO_STOCK = "APP_MONO_STOCK";
     public static final String APP_MODE_PAYMENTS = "APP_MODE_PAYMENTS";
